Update memcpy and memset function signatures
Changed the parameter types of __ESBMC_memcpy and __ESBMC_memset from
unsigned long long to size_t for better compatibility with standard C
library functions.

+++ src/c2goto/library/string.c
@@ -269,7 +269,7 @@ __ESBMC_HIDE:;
   return cpy;
 }
 
-void *__memcpy_impl(void *dst, const void *src, unsigned long long n)
+void *__memcpy_impl(void *dst, const void *src, size_t n)
 {
 __ESBMC_HIDE:;
   char *cdst = dst;
